Consolidate PDF report components under pdf-export.
