+++ assets/js/components/DashboardMainApp.js
@@ -89,9 +89,9 @@ import MetricsSelectionPanel from './KeyMetrics/MetricsSelectionPanel';
 import ModuleDashboardEffects from './ModuleDashboardEffects';
 import Notifications from './notifications/Notifications';
 import OfflineNotification from './notifications/OfflineNotification';
+import PDFDownloadButton from './pdf-export/PDFDownloadButton';
 import PDFExportRoot from './pdf-export/PDFExportRoot';
-import PDFDownloadButton from './pdf-generation/PDFDownloadButton';
-import PDFSectionsSelectionPanel from './pdf-generation/PDFSectionsSelectionPanel';
+import PDFSectionsSelectionPanel from './pdf-export/PDFSectionsSelectionPanel';
 import CurrentSurveyPortal from './surveys/CurrentSurveyPortal';
 import SurveyViewTrigger from './surveys/SurveyViewTrigger';
 import WelcomeModal from './WelcomeModal';
