+++ assets/js/components/pdf-export/PDFDownloadButton.test.tsx
@@ -19,7 +19,7 @@
 /**
  * Internal dependencies
  */
-import { PDF_DOWNLOAD_PANEL_OPENED_KEY } from '@/js/components/pdf-generation/constants';
+import { PDF_DOWNLOAD_PANEL_OPENED_KEY } from '@/js/components/pdf-export/constants';
 import { CORE_UI } from '@/js/googlesitekit/datastore/ui/constants';
 import { createTestRegistry, fireEvent, render } from '@tests/js/test-utils';
 import PDFDownloadButton from './PDFDownloadButton';
