+++ assets/js/components/pdf-export/PDFSectionsSelectionPanel/Header.tsx
