+++ assets/js/components/pdf-export/shared-react-pdf-components/PDFMetricTile.tsx
