+++ assets/sass/admin.scss
@@ -181,7 +181,7 @@
 @import "components/email-reporting/googlesitekit-invite-others-to-subscribe";
 
 // PDF Generation
-@import "components/pdf-generation/googlesitekit-pdf-download-panel";
+@import "components/pdf-export/googlesitekit-pdf-download-panel";
 
 // Setup
 @import "components/setup/googlesitekit-setup";
